fix: 카테고리 기능 수정

+++ src/main/java/com/sparta/spartaproject/domain/category/CategoryService.java
@@ -1,5 +1,6 @@
 package com.sparta.spartaproject.domain.category;
 
+import com.sparta.spartaproject.domain.CircularService;
 import com.sparta.spartaproject.dto.request.CreateCategoryRequestDto;
 import com.sparta.spartaproject.dto.request.UpdateCategoryRequestDto;
 import com.sparta.spartaproject.dto.response.CategoryDto;
@@ -8,10 +9,8 @@ import com.sparta.spartaproject.exception.ErrorCode;
 import com.sparta.spartaproject.mapper.CategoryMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.UUID;
@@ -21,6 +20,7 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class CategoryService {
     private final CategoryMapper categoryMapper;
+    private final CircularService circularService;
     private final CategoryRepository categoryRepository;
 
     @Transactional(readOnly = true)
@@ -40,7 +40,7 @@ public class CategoryService {
     @Transactional
     public void createCategory(CreateCategoryRequestDto request) {
         if (existsCategoryByName(request.name())) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "이미 존재하는 카테고리입니다.");
+            throw new BusinessException(ErrorCode.ALREADY_CATEGORY);
         }
 
         Category newCategory = categoryMapper.toCategory(request);
@@ -52,7 +52,7 @@ public class CategoryService {
     @Transactional
     public void updateCategory(UUID id, UpdateCategoryRequestDto update) {
         if (existsCategoryByName(update.name())) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "이미 존재하는 카테고리입니다.");
+            throw new BusinessException(ErrorCode.ALREADY_CATEGORY);
         }
 
         Category category = getCategoryById(id);
@@ -63,17 +63,18 @@ public class CategoryService {
 
     @Transactional
     public void deleteCategory(UUID id) {
-        // TODO: 사용중인 카테고리인지 검사 - 사용중이면 삭제 X
-
         Category category = getCategoryById(id);
-        categoryRepository.delete(category);
 
+        if (circularService.getStoreCategoryService().existsByCategory(category)) {
+            throw new BusinessException(ErrorCode.CATEGORY_IS_USED);
+        }
+
+        categoryRepository.delete(category);
         log.info("카테고리: {}, 삭제", id);
     }
 
 
     public Category getCategoryById(UUID id) {
-        // TODO: 카테고리 에러 메시지 적용, 404
         return categoryRepository.findById(id)
             .orElseThrow(() -> new BusinessException(ErrorCode.CATEGORY_NOT_FOUND));
     }
